AI_medical_projects/HEALTHWISE-HACKKU25/requirements.txt
streamlit==1.32.0
google-generativeai>=0.3.2
python-dotenv==1.0.1
typing-extensions==4.9.0
protobuf==4.25.3 